Fixed BFDR synthetic record generators
